Remove state.js
