+++ src/components/CardTemplate.jsx
@@ -2,6 +2,6 @@ import styled from 'styled-components'
 
 export default styled.div`
 	box-shadow: var(--elevation);
-	background-color: var(--white);
+	background-color: white;
 	border-radius: 4px;
 `
